webapp/requirements.txt
flask>=0.9
flask-sqlalchemy>=0.7.8
gevent-socketio>=0.3.5-rc2
nltk>=3.0
flask-login
flask-wtf
